- 단기예보 안정화 - 스케줄러와 시간 계산 로직 분리
